requirements.txt
aiohttp==3.4.4
async-timeout==3.0.1
attrs==18.2.0
cachetools==3.1.1
cffi==1.11.5
chardet==3.0.4
discord.py==1.1.1
gitdb2==2.0.5
GitPython==2.1.11
google-api-python-client==1.7.6
google-auth==1.6.1
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.2.0
html5lib==1.0.1
idna==2.8
lxml==4.2.5
matplotlib==3.0.2
markdownify==0.4.1
mpmath==1.1.0
multidict==4.5.2
nltk==3.4
pika==0.13.0
psycopg2==2.7.6.1
psycopg2-binary==2.7.6.1
pycparser==2.19
PyNaCl==1.3.0
singledispatch==3.4.0.3
six==1.12.0
smmap2==2.0.5
sympy==1.3
systemd-python==234
websockets==6.0
yarl==1.3.0
zalgotext==0.2.0
